% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyModels.R
\name{ApplyModels}
\alias{ApplyModels}
\title{Title Apply various ML models to your datasets and compare the results}
\usage{
ApplyModels(working_df, model_names = c("RF", "LDA", "NB", "SVM", "KNN"),
  class_column_index = -1, split_ratio = 0.66, save_results = TRUE,
  plot = TRUE)
}
\arguments{
\item{working_df}{input dataset for classification}

\item{model_names}{a list of models to be appied on the data}

\item{class_column_index}{the column index of the classes}

\item{split_ratio}{training to test ratio}

\item{save_results}{if set to true results are saved}

\item{plot}{if true plots confusion matrix}
}
\value{
accuracies a dataframe containing model names and accuracies
}
\description{
Title Apply various ML models to your datasets and compare the results
}
