% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyNN.R
\name{ApplyNN}
\alias{ApplyNN}
\title{Title}
\usage{
ApplyNN(raw_df, n_var)
}
\arguments{
\item{raw_df}{input data}

\item{n_var}{number of predictors. i, e. independent variables}
}
\value{
results
}
\description{
Title
}
\examples{

}
