% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{Impute}
\alias{Impute}
\title{Title}
\usage{
Impute(data, time_period_length = 20, k = 5)
}
\arguments{
\item{data}{The input dataset, must have x_axis, y_axis, z_axis columns}

\item{time_period_length}{The time period after of before the NA value}

\item{k}{Number of samples to be considered for NA imputation}
}
\value{
main_df Imputed results
}
\description{
Title
}
