% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateFeatures.R
\name{GenerateFeatures}
\alias{GenerateFeatures}
\title{Generate new features from raw data for a given time window}
\usage{
GenerateFeatures(raw_df, x_axis_column = 1, y_axis_column = 2,
  z_axis_column = 3, window_size_sec = 1, frequency = 30)
}
\arguments{
\item{raw_df}{a dataframe which has at least three coluns, x_axis, y_axis, z_axis}

\item{x_axis_column}{column number of x axis or}

\item{z_axis_column}{}

\item{window_size_sec}{windows size in second}

\item{frequency}{sampling frequency}
}
\value{
new_features a dataset contaning generated features
}
\description{
Generate new features from raw data for a given time window
}
\examples{

\dontrun{
# load a sample dataset
library(activityCounts)
load("sampleXYZ")
sampling_freq <- 100

# prepare the dataset by setting proper column names
raw_df <- sampleXYZ \%>\% rename("x_axis" = accelerometer_X,
 "y_axis" = accelerometer_Y, "z_axis" = accelerometer_Z)

# consider a one second window
window_size_sec <- 1

# generate new features
GenerateFeatures(raw_df = raw_df, window_size_sec = window_size_sec, frequency = sampling_freq)
}

}
